Automatically focus on article content if mouse hasn't moved
It's always annoyed me that, whenever I try to open Open2v2's changelog
and go to the bottom, I have to move my mouse first. And moving your
mouse is really unnecessary.

I thought that by default we would get the focus on the article content
for free if we haven't moved our mouse yet, but it turns out that if we
haven't moved our mouse yet on the window, both our X and Y position are
0, for some reason, instead of being something sane like nil. This seems
to be a limitation of SDL/LÖVE, but I haven't found much about this when
I searched on the internet. And if our X is 0, then we've focused on the
article list sidebar.

So, I've made it so that you will focus on the article content if your
mouse's X position and Y position both just happen to be 0. Yes, this
means if your position is legitimately 0,0 instead of just happening to
be unset, you will scroll the article content instead of the sidebar
where your mouse actually is. But I'm pretty sure that will never happen
and you'd have to know about this commit in order to intentionally do
that, anyway. And if it does happen, you'd go like "wtf" and reflexively
move your mouse and it'd stop being at 0,0 and it'd be fine anyway.