Add minimap preview under subtools
This shows what the minimap for the current room is, as well as showing
you which tile on the minimap you're modifying, if you're hovering over
one on the grid. Also, there's a nice border around it.

I've had to reduce the padding between subtools for tools 1 and 2, but
it's fine. Additionally, I've also made sure to update the Potato code
so that the mouseon still works. Luckily, no one reads these commits
other than Dav, and especially not Ally, so I can talk about the Potato
all I want.