Fix '/'-and-'?'-shortcut-no-dirty code not dirtying typing at all
Somehow before I made it so that the level doesn't get dirtied if you
type anything at all (but everything else, including backspacing /
deleting forwards, dirties it) with my other fix for the problems caused
by my earlier patch to fix '/'-and-'?'-shortcuts-dirty-the-level. So now
hopefully this is the end of the saga, and eventually we won't have to
deal with issues like this in the new dialog system.