Set count.startpoint when redoing start point placement
If you placed a start point, then undid it, count.startpoint would
become nil, as it should. But if you redid it, then it would still be
nil. Then Ved would let you place another start point, because it saw
that count.startpoint was nil, and thought there was no already-existing
start point based off of that.

Note that this is only if you're PLACING the start point, and not MOVING
the start point. If Ved correctly identifies that there already is a
start point in the level, it will simply move that start point to the
location your cursor is, instead of deleting it and making a new one.
This bug does not apply if you move a start point, and undoing/redoing
the relocation of a start point is perfectly fine.

This bug has existed in Ved ever since you could undo/redo placing
entities. Until now, where this commit fixes it.