Add an FC for multiple start points
There's basically no reason to ever have multiple start points, ever.
VVVVVV will just pick the first one in the level anyway. Unless it's
different on other versions, but I don't think so, and I'm not going to
bother to test other versions.

While I was fixing the bug where you could place multiple start points
by undoing and redoing the start point placement
(e08b63b95b32dfa1ffffde0532292fa2c340d9f5), I noticed that Ved didn't
seem to have an FC for multiple start points, so I added this.

Unfortunately, my code is a bit ugly. I just wanted to put a `continue`
when a duplicate start point was found, but this is Lua, with no
`continue` statement, so my first attempt was to use `goto`. It didn't
seem to work (also, my Lua linter complained, even though my linter is
wrong). So then I resorted to simply collecting up the IDs of the
duplicate start points, and when all the entities were parsed, simply
`table.remove()` every single last one of them.

`table.remove()` is a bit bad because it's really inefficient if you
have to do it a lot in quick succession (because every single time it's
called it updates the indices of every other element after the one you
remove - which is also why if you're going to systematically remove
things from a table in a loop, you have to iterate over the table
backwards), which is what would happen if someone made a level with a
giant amount of start points. So I assume no one will ever make a level
with 1,000 start points, unless they read this commit and decide to
troll me, but no one ever reads these commit messages other than Dav,
and Dav wouldn't troll me. Anyways, all of that is to say that the
inefficiency of `table.remove()` will never be an issue. I hope.
Probably.