Remove redundant 3rd arg from null FC langkeys
They're used to indicate the key of the thing in the table you passed to
the second argument upon which the plural calculation should be based.
They default to 1, and besides, I've only passed it a table with one
item in it. Shouldn't be that hard for it to figure out.