Make dialog scrolling mutually exclusive from non-dialog scroll
What I did is I made it so the dialog.is_open() branch is a part of the
same tree as the one immediately below it, and made sure it's at the
top. That way, if a dialog is open, there is no other way that you can
scroll anything else, i.e. the dialog takes priority.

This fixes a bug where you could scroll with Home/End while a dialog was
open, e.g. when in the script list and having the new-script-creation
dialog open.