Changed two shortcuts (Q and K in main editor)
Q (go to room) is being changed to G since that's a more logical key,
and to make room for Q and A scrolling a VVVVVV-CE tower in the future.
Outside of towers, I guess A is the altstate key since I think they
won't be a thing in towers (someone please confirm), and Q could
possibly be used for something that is unavailable in towers as well.
Going to a different room by coordinates is definitely something that
should work in towers.

K (show background) is now Shift+;. ; is the button that shows which
tiles are checked for solidity on the minimap. This makes room for one
more possible VCE-related shortcut.

[pre07]