Fixed crash when dropping a file onto the graphics viewer
Some variable names were confusing, and it turns out
love.directorydropped provides a path string and love.filedropped
provides a DroppedFile, so `path` was a misnomer. Whoops. The hook
won't change and will still be passed this DroppedFile, and not
suddenly be changed to pass the path as was previously suggested.

[pre10]