Refactored right click menu system
There's now a function rightclickmenu.mousepressed(x, y, button) which
is called from love.mousepressed and takes some mouse checking out of
the function that shows the options, and importantly there's now a
function rightclickmenu.handler(RCMreturn) which takes checking the
clicked option out of love.update, and instead makes it handled as a
direct result of clicking the button. Not perfect, but much better.

[pre11]