Preparing for changeable VVVVVV limits
There's now a lookup table `limit`. For now this is constant and set to
VVVVVV's limits (including, now, 100 trinkets and crewmates, which I'm
not yet sure about since the first version of VVVVVV that ensures
trinkets and crewmates above 20 are reset properly is yet to be
released). Lots of hardcoded limits are now changed to use this table,
but I'm sure I haven't gotten everything. I focused on map width and
height, trinket counts and crewmate counts.

This will all make it easier to add VCE support.

[pre13]