Added F shortcut to plugin error screen
Press F to  open your plugins folder, so you can fix or remove the
plugin which causes the orange screen. Came up with this idea because
everyone who has installed Temp Fix Tiles to fix the tileset
transparency bug 1.7.0 had, will get an orange screen after downloading
1.7.1 because the code that the plugin fixes doesn't exist anymore. Now
it'll be easier to remove that plugin. (And to tell people how to do
it)

[pre16]