Removed simulation of buggy translucent pixel tile drawing
Since 1.7.0, tiles would be drawn more consistently with how VVVVVV
does it, pixels that are not fully opaque or transparent would show up
as some kind of blue color. This has been fixed in VVVVVV's repository
now (https://github.com/TerryCavanagh/VVVVVV/pull/103), so we have more
reason to do away with it. Also because the bug that happened in LÖVE
0.10.x and lower has recurred, so might as well get rid of it right
away.

[pre17]