Worked further on VCE support
loadlevel and createblanklevel now return extra, which is a table where
all new non-VVVVVV-target things are stored. limit is now also returned
by loadlevel and createblanklevel, and correctly adapts to VCE levels.

Now altstates, towers, teleporters and time trials are recognized if
they're empty (self-closing tags) and if they're not empty, a failed
check is generated, to warn that that data will be lost. That means,
once handling for new properties of levelmetadata and entities is
added, we can start to say that Ved really supports loading and saving
VCE levels without data loss as long as there's no altstates, towers,
teleporters or time trials in there. Then one by one these can be
loaded in.

Saving levelmetadata in VCE style now works as well.

[pre18]