Moved metadata entity from (20,20) to (100,100) 0-indexed
This applies to both VCE and VVVVVV levels.
This of course introduces a backwards incompatibility when newly saved
levels are opened in earlier versions of Ved, but this is definitely
outweighed by the possible problems when opening a level in VCE,
enlarging the level size, and suddenly finding a roomtext with tons of
dollar signs in one of your newly acquired rooms. When doing that with
levels which have the MDE at its old position, just open in a new
version of Ved and save; in VVVVVV levels it'll look for the MDE in
either location, but it'll always save in (100,100) from now on. In VCE
levels, only (100,100) is considered. If you ended up with a VCE level
with an MDE in (20,20), just use the entity properties dialog to change
its position to 4000,3000, then save and load, and it should start
working again and be out of your way.

The MDE format will still be updated to allow for 1000 flag names.
Without 999 dollar signs if not necessary.

[pre22]