Remove remnants of libvloader
We're no longer using it, so we no longer need to mention its name in
comments, nor have the lib laying around.

This also enables playtesting on macOS, although I haven't tested macOS
to see if it works, but it should. I also think that Windows should work
as well, although I haven't tested that, either, but macOS and Linux are
similar enough that I'm making the call and immediately enabling macOS
playtesting.