Made language screen not crash if language names aren't UTF-8
So, got a report that the language screen was crashing, turned out the
name of the Russian language file was being passed in Shift-JIS for
some reason. I'll later move to a slightly different system where
language files use codes as their names and where the name is displayed
based on that, but for now the temporary fix is, if the string causes
an error due to not being valid UTF-8, it's replaced by a question
mark.

[pre31]