Made metadata entity flag name storage have a dynamic length
Flag names were previously serialized as a list of 100 possibly empty
names. Now the list will only be made until the highest filled flag
name, and it can go above 100, for VCE purposes.

Increasing the metadata version isn't necessary imo since the metadata
entity is at a different location from the next version of Ved onward,
in other words, versions of Ved incompatible with this change already
don't see the metadata with an increased version anyway, so they can't
handle that anyway. But feel free to call this version 3A or something.

[pre34]