Restored compatibility of the "outdated LÖVE" screen with LÖVE 0.7.x
Thought it'd be a bit professional to have the error screen for
outdated versions of LÖVE work properly in more than one or two
outdated versions of LÖVE, like it had before. How did people check the
version number of LÖVE back then? Apparently you just didn't, and just
had the target version of LÖVE in your conf file, so that part of the
error screen can only report it's 0.7.x or older.

And never mind that the error screen looks like that:
https://i.imgur.com/uKiFcwj.png

2011 LÖVE didn't yet have UTF-8 support, but I think most people will
get the message anyway.

[pre40]