Display script box placement order when hovering over a script box
By default, the placement order is displayed to the left of the script
name at 1x scale. If the script box is too far left, it gets moved to
just above the script name. If the script box is too far left AND too
far top, it gets moved to just below the script name.

This closes only a fraction of #79.