Changed language files to have code as filename instead of native name
Instead of being named like Français.lua or Русский.lua, all languages
are now named as their language codes, such as fr.lua and ru.lua.

First off, that's a better system. If a language name ever needs to be
changed (maybe you want something like "English (UK)" to be renamed to
"English, UK") then people who used that language won't have their
setting reset if you don't want to add a mapping for old to new names.

Second, it was a good sign for the state of Unicode handling that the
former system hasn't given any problems for years - until last month,
when two people got problems with it:

- One used WinRAR to make a .love. Apparently WinRAR malforms filenames
  in code page 437 or 850 or something related whenever it can. No, not
  even ISO-8859; it used an OEM code page. "Русский" in the filename
  was encoded in UTF-8 just fine, "Français" and "Español" in filenames
  were encoded in, ugh
- One had the language screen crash because the Russian language name
  was somehow encoded in Shift-JIS, while the others were in UTF-8. I
  don't know if a zipping program was used there or not, but either way
  using only ASCII for filenames in Ved would solve this issue.

Now there's an info.xml file next to the language files which holds all
the languages' names, which could easily be extended for further
language properties in the future, like using a different font or
changing text direction. The language setting is now stored using the
code as well, but existing language settings using the full language
name are still recognized, based on the names in the info file.

[pre45]