Fix entities being able to be placed on flip tokens and teleporters
Apparently, there's a table in `const.lua` called `entitytoolbyid` that prevents entities to be placed on the
entities in the table. It probably does other stuff too, as I'd assume all entities should be treated the same
no matter whether they're invalid or not, but for our purposes, adding the new entities to the table seems to
fix the problem.