Return XML buffer if path is nil
This lets us get at the actual contents of the level file, but without
actually writing it to disk as well.

This is in preparation for the reworked playtesting interface with
VVVVVVV: Community Edition (and soon VVVVVV 2.3) where we simply pipe
the level into STDIN and never have to bother with the filesystem as a
middleman (which, due to portability, the original proposed change to
upstream VVVVVV was rejected).