Added (half-temporary) method of changing custom tileset/spritesheet
Pressing Ctrl+F9 will bring up a dialog where the custom tileset and
spritesheet can be chosen for that room. I say half-temporary, because
there's a bit of a dilemma here with the pre-existing Ved shortcut F9
to reveal other shortcuts, that was decided to be used for this in
VVVVVV-CE.

I don't know what would be a better shortcut other than F12, which
would only be marginally easier to remember only because it's at the
end of the function keys instead of in the middle, and just as much at
risk of being used by VVVVVV-CE as well. Moving F9 to F12 just wouldn't
be worth it. So I'm thinking of just not having keys to cycle between
custom tilesets/tilesheets, at least for now. Later there will be a
visible button to open this dialog.

[pre09]