Moved location where all input is paused after a state change
It was after the call to loadstate, which meant creating a new input
as part of moving to a new state would immediately pause that input.

Also added drawcas to script editor. The new input system now gives a
sign of life in the script editor! And now it's time to fix all the
bugs and remnants of the old system that cause conflicts here.

(cherry picked from commit 383b990a5a9ac2735890070a3653672647f7b558)