Add second return value, "what" to get_parent_path()
get_parent_path() only returns the "where", now we need it to return the
"what". This is only because I'm going to be adding file_exists() the
file funcs, and for filefunc_lin_fallback.lua, it will simply re-use its
existing directory_exists(), which needs to give the "where" and the
"what" separately instead of having both already combined with a
directory separator. However, file_exists() will be taking in an
already-combined path, so it will thus need to separate the path into
the "where" and the "what". But get_parent_path() only returns the
"where", so we need it to return the "what".

I do two string substitutions here to get around an edge-case that could
happen if I removed the second substitution and did the first
substitution starting at `-last_dirsep+1` instead of `-last_dirsep`. If
the path passed ends in a trailing directory separator, the "what" will
end up being the entire input passed (including the directory
separator), due to the fact that if the string ends in a directory
separator, `-last_dirsep+1` will end up being 0, which means it's no
longer negative, which means Lua will end up starting from the beginning
of the string instead of starting from a place from the end of the
string. So to get around this, I simply substitute starting at
`-last_dirsep`, which will never be 0, and then do a second substitution
to remove the directory separator afterwards.