Add file func file_exists()
Instead of taking a "where" and a "what", it takes in a full path.

For the Linux-macOS filefuncs, this is as easy as passing the path to
the C lib. For the Windows filefunc, we just copy the existing
directory_exists() and negate the is-a-directory check to explicitly
check that it's a file. For the Linux-fallback filefunc, we just re-use
directory_exists() since apparently directory_exists() just checks if a
path exists, not necessarily that that path is a directory. For the LÖVE
filefunc, we use love.filesystem.isFile(), however on LÖVE 11 and later
we have to use love.filesystem.getInfo() to check if the path has any
attributes of the type `file`.