Change VVVVVV path system to select file instead of directory
In VVVVVV 2.2 and below, it mattered which directory you ran the game
from. You would have to run the game from the folder CONTAINING the
folder 'x86_64/' or 'osx/', otherwise strange behaviors would result,
like the game segfaulting if a custom garphics file was missing, even
though it doesn't usually do that.

In VVVVVV 2.3 (including VVVVVV-CE, which I'll add separate path system
support for shortly), this issue has been fixed and you can run the
VVVVVV binary independently of any directory structure.

So there's no longer any need for this system where we ask users to
point to a directory. Instead, we can just tell them to point to the
file directly.