Add VVVVVV-CE support
There is now a separate config option, s.vvvvvvce, to give the path to
the VVVVVV-CE executable. This executable will be used if the level is a
VCE level, instead of the vanilla one.

Annoyingly, there's currently an issue on VCE's side (which to be fair,
I *could* work around, but it's really stupid if I have to and I'd
rather it just be fixed on VCE's end), where VCE still looks for
data.zip in the directory you run the binary from, NOT in the directory
the binary is located. Thus, VCE as of right now is unable to locate
data.zip and thus run properly in Ved playtesting, which is dumb.

I've no idea why VCE does this, as vanilla does NOT do this (as of 2.3),
but I'm pre-emptively blaming Leo because I'm pretty sure he messed with
some filesystem stuff and broke this. So good job, Leo.