Enable playtesting on Windows
Playtesting on Windows seems to actually work great! Except for the fact
that on my installation, a command prompt window seems to pop up and
stay up for the duration of the playtesting. Which we'll have to fix
Later(tm).

So I've gotten rid of the whole "separate playtesting_execute()
functions" thing, mostly because at this point, any OS-specific
differences are minor enough to only warrant an if-conditional inside
the grand unified playtesting_execute() function. (And to fix the
Windows command prompt thing, we'll have to modify playtestthread.lua
anyway, playtesting_execute() isn't where the problem is.)