Added fast path to roomdata_get for rooms within size limits
I thought I already added or proposed this? Either way, in like 100% of
the cases (rounded down, not up) only rooms within limits are shown,
and it's of no use to do a table.copy on the roomdata (which is used in
the other 0% of cases (rounded up) to determine distortion) so we just
return the roomdata directly if the room is within limits.

[pre21]