Added level conversion between VVVVVV and VVVVVV-CE
You no longer need to open a pre-existing VVVVVV-CE level to create a
VVVVVV-CE level, or open and save in VVVVVV-CE/VVVVVV or modify the XML
to change a level's target platform to one or the other. Converting a
VVVVVV level to VVVVVV-CE will allow VVVVVV-CE features to be used.

Converting a VVVVVV-CE level to VVVVVV will remove VVVVVV-CE features
from the level, and will remove all altstates, towers, VCE-only
entities, etc. It will limit the level to 20x20. But some things that
have already never been restricted in VVVVVV mode (like the amount of
scripts) are still the user's responsibility.

Also, VCE-only entities are now only shown in VCE mode (in V mode,
they're the red [E]). This is important because the raw entity
properties dialog allows people to change types manually, and it would
be misleading to display VCE-only entities as though they work in
VVVVVV levels.

[pre26]