Ported playtesting VVVVVV starting to Windows API on Windows
This allows us to start VVVVVV without showing a command prompt window.
Also, it probably prevents potential issues with character encodings in
the path to VVVVVV.

Also, since you can specify a working directory in CreateProcess, I
used that to work around an issue in VVVVVV-CE 1.0-pre1, which expects
data.zip to be in the current working directory, instead of next to its
own executable. 1.0-pre2 isn't out yet, so lots of people would
otherwise get that issue.

[pre31]