Increased robustness of XML parser for self-closing tags without space
VVVVVV 2.3 upgraded from TinyXML 1 to TinyXML 2. This means an empty
edEntities tag (the only tag that VVVVVV normally makes self-closing)
is now no longer stored as <edEntities />, but as <edEntities/>. That
turns out to be no problem - and has been no problem for a very long
time (thanks self for adding the extra check somewhere during the alpha
or beta stage!) ...

... but I recently added parsing of some self-closing VVVVVV-CE tags,
which were only expected to be self-closed _with_ the space.

I now made all those tags be recognized whether they have the space or
not. This was only really a problem for teleporter tags, now that I
think of it, since that's the only self-closing tag that actually
carries data instead of simply being empty anyway and thus be treated
as if it wasn't in the level file anyway.

[pre04]