Added support for 2.3/VCE level-specific assets folder
Well, there are different ways to include level-specific assets in the
levels folder, including a levelname.data.zip, but right now Ved
supports the method where there's a folder with the name of the level
in the levels folder. Level-specific assets should now be correctly
loaded and displayed when opening a level.

[pre06]