Added support for level-specific assets when playtesting
When playtesting, the level contents are piped to VVVVVV, so VVVVVV
doesn't know what the name of the level is, if it is saved at all.
Therefore it also can't load custom assets specific to that level.
To resolve that, at least in pull request 310 on the main VVVVVV repo,
support was added for a -playassets argument. Now Ved passes this
argument.

[pre07]