Fixed map display of trinkets and crewmates when shifting map right
Trinkets and crewmates that were in rooms in the leftmost column would
get visually smeared over the map, trinkets and crewmates that were in
the rightmost column would visually fall off the map instead of
wrapping around. Luckily, entities did not actually get duplicated or
removed, the indications on the map were just getting updated wrong.
This is now fixed.

[pre09]