Removed "FATAL ERROR" state and need for setting statecaught in plugins
It has proven to be pretty unnecessary for Ved to have a way to stop
working of its own accord and display an error message (crashes and
unsupported LÖVE screens don't count as such). It was only used to
indicate that the state number was not recognized, which just prevented
a black screen, but why would that ever happen and be important enough
to give an error for? Plugins that made their own states would need to
set the global variable statecaught to true in order to not have the
screen show up. And the message even referred to Ved as a game! I guess
I copy-pasted the initial structure from a game I was working on at the
time...

If I'd need a special "FATAL ERROR" message in the future, I'd probably
use the dialog system (like how there's a dialog on a black screen if
the settings file is corrupted and needs to be cleared)

[pre02]