Fixed lag/freeze when adding a textbox with extreme amounts of lines
This started in 1.7.1, because a line was added that indicates the
right border of a textbox. Since the longest line of text in the
textbox needs to be known to be able to draw that line, we could even
continuously try to find out what the longest line was in a text box
claimed to be 99999999 lines long... Regardless of whether the script
even had that many lines. Now it just stops iterating lines when the
end of the script is reached.

[pre05]