Made non-ASCII chars typable throughout Ved, obsoleting s.acceptutf8
Typing non-ASCII/Unicode characters used to be blocked in most of Ved
unless the hidden setting s.acceptutf8 was enabled, because Ved would
display UTF-8 correctly (in most cases) and VVVVVV wouldn't. At best,
VVVVVV would display question marks for the raw number of bytes, and at
worst, older versions of VVVVVV would immediately crash. What times
those were.

Now the upcoming version of VVVVVV adds support for displaying UTF-8,
as long as a font is used that has more characters than just ASCII, but
otherwise it'll display question marks for the number of codepoints,
not bytes. Not sure if VVVVVV 2.3 will be shipped with a non-ASCII
font (hopefully so), but VVVVVV-CE has it built-in, which we also
support. Ved's blocking non-ASCII characters is also based on a certain
short list of state numbers, and you were never blocked from pasting
them in anyway, so we now no longer try to block it.

[pre06]