Fixed state 21 (overlapping entities) and migrated it to new UI system
The function to check overlapping entities was broken in a way that can
only be described as having happened even later than the current time
of making this commit, 03:22.

It's still displaying room coordinates in a non-integer fashion but I'm
not gonna figure that our right now, at least now it displays something
useful.

[pre10]