Removed states 22/23 (script file load) and PleaseDo3DSHandlingThanks
State 22 allowed you to open an external script file in 3DS format with
lines separated with $ signs into the script editor, state 23 allowed
you to open an external script file with lines separated by simple
newlines.

I've decided to remove this for various reasons:
- It's extremely hidden, only accessible through state numbers
- It's user-unfriendly, go type in or copy-paste a full path without
  autocomplete or file listing
- It's pretty broken, unmaintained and unpolished
- It adds code complexity (particularly conditions inside conditions to
  see if the newline character is | or $)
- "PleaseDo3DSHandlingThanks"
- Would need to be migrated to new input system despite all these
  reasons

If loading a script from a file were to be supported again in the
future, it would need to be rewritten anyway. Now a fair amount of code
could be simplified.

[pre11]