Fixed exiting syntax highlighting options also exiting Ved options
This time the problem wasn't that one of the states wasn't migrated to
the new UI system yet - the problem was that LÖVE callbacks delegate to
both "Elements" and a generic UI-specific callback. The syntax
highlighting options checks for pressing Esc in the generic callback,
changes the state, and then immediately afterwards, the *new state's*
elements also get a chance to handle the event, so the OK button in Ved
options was also triggered.

Now, for all callbacks, the element callbacks only get called if the
state didn't change because of the generic callback.

[pre13]