Added state number cheatsheet to state 0, fixed search UI name
I've been wanting a list of state numbers in state 0 for a while, and
with the UI system that's easier to do without keeping around another
copy of the list that needs to be kept updated. Of course the list
still lacks states which haven't been migrated to the new UI system,
but that's gonna be done soon.

Turns out I forgot to change the TEMPLATE name in the search state, so
that's now fixed.

[pre15]