Add white and orange text box colors
VVVVVV 2.3 adds being able to use the white and orange colors in the
text() command.

Alt text box colors have been updated accordingly, because you just need
to set the red channel to either 0, 13, 14, 15, 16, 19, or 20 (the
actual precise value depends on which color you pass in to
createcrewman(), but the variation is so little that it really doesn't
matter).