Migrated state -2 (init/tostate 6) to new UI system
This also has state-specific code in love.update so I thought this one
shouldn't be skipped either.

So, the only state that still needs to be migrated to the new GUI
framework is state 1, the main editor. Decided to do this last because
this would have the most code scattered all around the LÖVE callbacks.
And then we can finally tear down some elseif chains!

[pre20]