Outline room names to be consistent with VVVVVV 2.3
In VVVVVV 2.3, the translucent room name option was added, and if the
room name is translucent then the font gets an outline. Ved's room names
are translucent by default, but they do not have an outline. So to be
consistent with vanilla, we should outline room names as well.

Room name editing and messages that use the room name are also outlined
as well.

The love.graphics.setScissor() had to be updated as otherwise it would
cut off the topmost and bottommost rows of outline pixels.