Added immediate flushing when saving level files on Windows
Sometimes (I think once or twice in the last years, plus once more
yesterday) someone reports that a level file has turned into
megabyte(s) of NUL bytes. A mystery, even when considering that in both
cases I could find, the system had also suddenly restarted or BSOD'd.

It seems like the problem is that Windows waits before actually writing
files to disk for an unknown period of time (in my testing about
6 seconds, but this might be longer in other cases?) meaning if the
power is suddenly lost while the file is still buffered, it's lost.

Luckily there's a Windows API function to flush changes to a file
immediately, so that's now done before closing it.

I have no idea if Windows waited unreasonably long (minutes?) in the
cases where this happened, and it's never going to be for certain that
these reports were caused by this exact problem and not something else,
but it's a big smoking gun that 5 seconds after a file is successfully
saved, you can cut the power and the file contents turn out to be
nulled. Of course a BSOD at just the wrong moment can always ruin data,
but now it has to happen before Ved reports the file was saved
successfully. And I'd think that's in line with most people's
expectations, that if a file is "saved successfully" it's on disk, but
if a BSOD happens in the middle of a save, that file will be corrupted.

[pre21]