Migrated state 1 (main editor) to new UI system
I did this slowly over the course of... Damn, 2 weeks almost. But I
was also busy planning and executing webserver migrations, along with
other stuff; and this was easily the most complicated state to migrate
(and refactor). Hence why I saved it for last, so there'd be as little
code belonging to other states as possible in the middle of the main
editor's shortcuts and all.

Especially the keypressed code is now more cleaned up than it was, with
a lot of the conditionals being stripped of their repeated instances of
`and not editingroomname and editingroomtext ~= 0`, `and nodialog`, and
of course, `and state == 1`.

Fixes #118.

[pre22]