Added Enter-Enter hotkey to playtest at first checkpoint or startpoint
Now you don't have to move your mouse to the place you want to start
playtesting, as long as you'd be happy starting wherever VVVVVV's
editor would place you when pressing Enter. If you would be, just press
Enter a second time and it places you at the startpoint if it's in the
current room, otherwise at the earliest placed checkpoint, and if the
room has no checkpoints, you'll just start in the center of the room.

[pre24]