Change version number to 1.8.3
Also, I'm switching to the imperative form for my commit titles from
now on. For the Ved repo I've always been wanting to use messages that
are easier to understand and "friendlier" for non-programmers, which
meant "Added this feature" would be better than "Add this feature". But
honestly, this is basically the only repo I commit to where that's
still the (semi-unwritten) rule, and imperative titles have gotten more
natural for me than they used to be. I guess they're probably not too
hard to read for others either.

[pre01]