Change fully textual warnings in script editor to "warning lights"
Don't worry, you don't have to learn the meaning of new icons - on the
contrary, if you hover your mouse over an icon, there'll be an
explanation what "Direct script box/terminal reference!" or "Load
script required!" actually means.

This frees space in the script editor, prevents new elements in the
script editor from potentially overlapping with others, and allows
adding more warning messages than the current two because there's a lot
of room for them, and they don't have to be tersely explained or
abbreviated. This is easier to get right for translators, too, this has
been a place where translations needed changes several times.

https://cdn.discordapp.com/attachments/403718283855134720/771189389040549931/unknown.png

Also, language files have been synced to get the new strings in every
language, so all the language files are now up-to-date with all changes
by translators.

[pre03]