Change setlocale LC_ALL in Linux/Mac filefunc library to LC_MESSAGES
The locale is set so operating system error messages in file listings
are displayed in the user's language. But LC_ALL is a bit much - if you
select 12-hour time in a non-English system locale, you might not see
"am" or "pm" because of this. It vaguely remember it being possible to
mess with stuff like tonumber as well, but we only need the messages.

[pre03]