Fix ved_require() disregarding package.loaded when loading source edits
There are three branches to ved_require(). The first two use require(),
which sets package.loaded automatically, so there's no risk of
double-loading a module there. However, the third doesn't use require(),
and instead loads the file manually - which is fine, except for the fact
that it neither checks nor sets package.loaded. Meaning, it's possible
for a double-include to occur, and then bugs would result from that.

I've made it so that it checks for package.loaded, to prevent it from
loading at all if it's not already set as loaded, as well as sets
package.loaded if it loads the module successfully.