De-duplicate filefunc_linmac version number
This will make it so there is only one place for each OS that you have
to change in order to bump the version number of the filefunc lib.
Previously, it was duplicated all over the place, and you would have to
make sure to change it in multiple places.

My project format implementation (!43) is definitely going to require
some filefunc changes, but I'll let Dav handle actually changing the
version number (presumably he'll do it when the next version of Ved gets
released).