Don't put closing edentity tag on next line
This makes it so that the closing </edentity> tag will be put on the
same line as its opening tag. This is because putting it on the next
line and having the tag on the same indentation as its opening tag will
technically result in adding a newline plus 12 space characters into the
contents of the tag, and parsing this as intended requires a modified or
non-standard XML parser to do so.

I know that 2.2 and below place the closing tag on the next line, but
2.3 removes this because it was annoying to maintain that specific XML
output when its XML code got fiddled with. In particular, ever since it
upgraded to TinyXML-2 (from TinyXML-1), 2.3 currently uses a hack that
specifically strips out a newline plus 12 space characters from the end
of the tag's contents if it specifically has those characters, in order
to maintain backwards compatibility with previous XML files that placed
the edentity tag on the next line.

I'd like to encourage XML files to keep the edentity tag on the same
line as the opening one, to minimize the need for hacky parsing like the
above, and since a lot of people use Ved, I want Ved to do this same
behavior, too.