Remove VCE tools' ^1/^2/^3 hotkeys to fix re-pressing tool hotkey
You could always press a certain tool's shortcut while it was already
selected to switch back to its first subtool, but that has been broken
for a while. I want to overhaul the entire tools system at some point
soon, but this is an annoying enough bug that I just decided to fix it
by removing the shortcuts for the VCE flip token, coin and teleporter
tools, Shift+1/2/3. VCE has become obsolete so those shortcuts are not
as needed, and in my current drafts for the new tool system I'm also
thinking it'd be way simpler not to have to support hotkeys with
modifier keys in them anyway.

[pre07]