Fix audio duration bar sometimes going out of bounds
If a song/sound is started close to the end, LÖVE may fail to reset
the time to 0 when looping. Which means our time indicator and bar,
which just report what LÖVE is reporting, will go out of bounds.
Now it simply modulos the elapsed time with the total duration to
avoid this.

[pre08]