Remove debug tileset creator
Pressing Lctrl+\ in debug mode got you the tool I used in 2015 for
adding all the tilesets and tilecols into Ved. Well there's two
versions, a slow one (click 30 times for each tool) and a shortcut one
(click once for each half of the tiles picker or tile 0). And I even
still used it again this year to add the Tower tileset for VVVVVV-CE.
But it's a bit of a technical debt, and I don't think it will be needed
anymore, so I'm removing it. Worst case we can go back in git history
and temporarily use it again.

[pre08]